6972305 Preferred DC not selected after setting pdc via sharectl 6971047 smbd hang during FVT regression test 6711195 Sparc:Get error "Windows Explorer has stopped working" once click on Details fr Properties on Vista PSARC/2009/464 Offline attribute 6972515 Offline attribute - PSARC 2009/464 PSARC/2010/037 Windows Sparse Attribute 6972519 Windows Sparse Attribute - PSARC 2010/037 6719444 [CLI] - idmap help's output is not consistent with man pages's 6975449 idmap test suite needs idmap_cache_get_data() from libidmap 6974351 OpenSSL still taking smbd down
--HG--
rename : usr/src/lib/libidmap/common/idmap_priv.h => usr/src/cmd/idmap/idmap/namemaps.h

+++ usr/src/lib/smbsrv/libsmbns/common/smbns_browser.c
@@ -19,8 +19,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2010 Sun Microsystems, Inc.  All rights reserved.
- * Use is subject to license terms.
+ * Copyright (c) 2007, 2010, Oracle and/or its affiliates. All rights reserved.
  */
 
 #include <sys/tzfile.h>
@@ -834,8 +833,8 @@ smb_browser_send_HostAnnouncement(smb_hostinfo_t *hinfo,
 	    ++hinfo->hi_updatecnt,
 	    next_announcement * 60000,	/* Periodicity in MilliSeconds */
 	    hinfo->hi_nbname,
-	    hinfo->hi_version.sv_major,
-	    hinfo->hi_version.sv_minor,
+	    (uint8_t)hinfo->hi_version.sv_major,
+	    (uint8_t)hinfo->hi_version.sv_minor,
 	    type,
 	    SMB_SERVER_SIGNATURE,
 	    hinfo->hi_nic.nic_cmnt);
